Update service dependencies to start on early boot
This changes log2ram dependencies to start as early as possible, before other
services that may require access to /var/log, including rsyslog and
systemd-journald.

Additional dependencies on shutdown.target are required to ensure that
the service will be stopped at shutdown/reboot when DefaultDependencies
is set to "no" [1].

RequiresMountsFor declares dependency on /var mount when it is a mount point.

IgnoreOnIsolate prevents stopping the service when isolating another unit.

[1]: https://www.freedesktop.org/software/systemd/man/systemd.service.html