Update README.md
Log2Ram handles the detection of rsync automatically. If rsync is available on your system, it will be used for synchronization; otherwise, the script will fall back to using cp.

The documentation has been updated to reflect that `USE_RSYNC` is no longer required in the config file (/etc/log2ram.conf).